+++ src/capture.cpp
@@ -9,7 +9,7 @@ namespace cell_world {
             cool_down(parameters.cool_down){
     }
 
-    bool Capture::is_captured(const Location &predator_location, double predator_theta, const Location &prey_location) {
+    bool Capture::is_captured(const Location &predator_location, float predator_theta, const Location &prey_location) {
         if (!cool_down.time_out()) return false;
         if (predator_location.dist(prey_location) > dist_max) return false;
         if (angle_difference(predator_location.atan(prey_location),predator_theta) > theta_diff_max) return false;
